Filtering, deleting single students/teachers added
Filtering is now possible on the main  activity, additionally, teachers and students are now able to be deleted from within the activities. Still to do is implement foreign keys so that if a Teacher is deleted the subsequent students are deleted or it throws an error that the students must be reassigned. Additionally to do, make it possible to edit student information.